+++ NEWS.md
@@ -1,8 +1,9 @@
 * small modifications to the translation (#4)
+* added (ita) module for focus on Piemonte (20200318)
 
 # covid19ita 0.12.1
 
-* added module for focus in friuli (20200318)
+* added (ita) module for focus on friuli (20200318)
 * data update (after dpc fix)
 
 # covid19ita 0.12.0